Changed parameter of destroy function to take void pointer
This is to avoid having to cast the free function to `(void(*)(void*))`
when ever we use it as an argument. This is a PITA when using the g++
compiler, as it is stricter. Also I fixed some warnings from the g++
compiler in the same commit.

Ticket: None
Changelog: None
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>